Late binding of default setTimeout and clearTimeout
If `setTimeout` and `clearTimeout` aren't explicitly set in the `createDelay()` call, bind their values when the promise is created. This allows `sinon.useFakeTimers()` to work while still allowing other implementations of those functions to be passed to `createDelay()`. Previously the default `setTimeout` and `clearTimeout` were set aside when the module was loaded, meaning default `delay` and `delay.reject` didn't notice if they were later stubbed out.